Extract pool
