+++ crates/cubecl-runtime/src/stream/scheduler.rs
@@ -4,9 +4,5 @@ pub trait SchedulerStreamBackend: 'static {
     fn execute(&mut self, tasks: impl Iterator<Item = Self::Task>);
 }
 
-/// Manages multiple streams with synchronization logic based on shared bindings.
-///
-/// This struct handles the creation and alignment of streams to ensure proper synchronization
-/// when bindings (e.g., buffers) are shared across different streams.
 #[derive(Debug)]
-pub struct SchedulerMultiStream<B: SchedulerStreamBackend> {}
+pub struct SchedulerMultiStream {}
